Extract LocaleInfo into a seperate file
