+++ src/pages/_app.tsx
@@ -1,7 +1,5 @@
 import React from "react";
 import { AppProps } from "next/app";
-import type { IconProps } from "@radix-ui/react-icons/dist/types.d";
-import { ArrowLeftIcon, ArrowRightIcon } from "@radix-ui/react-icons";
 import { translateFunc } from "../locales/translate";
 import { useRouter } from "next/router";
 import { ILayoutContext } from "../interfaces/layout-context";
@@ -9,6 +7,7 @@ import { ReactLayoutContext } from "../contexts/layout-context";
 import { QueryContext, ReactQueryContext } from "../contexts/query-context";
 import { PageContext, ReactPageContext } from "../contexts/page-context";
 import { DynamicContentServer } from "../lib/dynamic-content-server";
+import { getLocaleInfo } from "../locales/locale-info";
 import {
 	getPathData,
 	getSiteTitle,
@@ -27,30 +26,7 @@ function App({ Component, pageProps }: AppProps) {
 	const layoutContext: ILayoutContext = {
 		locale,
 		compLocale: pathData?.locale,
-		localeInfo:
-			locale === "he"
-				? {
-						direction: "rtl",
-						right: "left",
-						left: "right",
-						arrowLeft: (props: Partial<IconProps>) => (
-							<ArrowRightIcon {...props} />
-						),
-						arrowRight: (props: Partial<IconProps>) => (
-							<ArrowLeftIcon {...props} />
-						),
-				  }
-				: {
-						direction: "ltr",
-						right: "right",
-						left: "left",
-						arrowLeft: (props: Partial<IconProps>) => (
-							<ArrowLeftIcon {...props} />
-						),
-						arrowRight: (props: Partial<IconProps>) => (
-							<ArrowRightIcon {...props} />
-						),
-				  },
+		localeInfo: getLocaleInfo(locale),
 		translate,
 		getSiteTitle,
 		getSiteSubtitle,
